EEN1037-Group-Project-main/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django_extensions
dj-database-url